+++ src/pages/StudyFormPage/StudyForm.jsx
@@ -1,16 +1,8 @@
 import React, { useEffect, useState } from "react";
 import styles from "./StudyForm.module.css";
-import Button from "../../components/Button/Button";
-import { createStudy, updateStudy } from "../../api/studies";
-import Toast from "../../components/Toast/Toast";
-import useToast from "../../hooks/useToast";
-
-//이미지는 이렇게 가져와야 정상적으로 작동할 확률이 높다.
-import desk1 from "../../assets/images/desk1.jpg";
-import desk2 from "../../assets/images/desk2.jpg";
-import tile from "../../assets/images/tile.jpg";
-import leaf from "../../assets/images/leaf.jpg";
 import { useLocation, useNavigate } from "react-router-dom";
+import Form from "./components/Form/Form";
+import Title from "./components/Title/Title";
 
 //type
 //- modify : 수정을 위한 ui,
@@ -20,317 +12,23 @@ import { useLocation, useNavigate } from "react-router-dom";
 function StudyForm({ type = "create", study = {} }) {
   //일단 지금은 navigate로만 접근하기 때문에, 실질적으로 props는 필요가 없음. 그러나 혹시 모르니 남겨둔다..
   //만약 다른 곳에서 얘를 컴포넌트로 사용하게 된다면..그러면 필요한..
+
+  //현재 Form의 타입과 주어진 study를 검사해보기 위함.
   useEffect(() => {
     console.log("type=>", type);
     console.log("study=>", study);
   }, []);
 
+  //navigate로 넘어왔을 시, location으로 인자를 꺼내준다.
   const location = useLocation();
   type = location.state?.type || "create"; //없을 경우, "create"가 유지되는 것이 아니라, undefined가 들어가게 되는 것 같다. 여기에도 그냥 "create"를 default값으로 넣어주었다. props의 create의 기본값은 필요없어지지만, 이해도를 위해 남겨두겠다.
   study = location.state?.study;
 
-  const navigate = useNavigate();
-  //input 상태 변수들
-  const [nickname, setNickname] = useState(study?.nickname || "");
-  const [title, setTitle] = useState(study?.title || "");
-  const [description, setDescription] = useState(study?.description || "");
-  const [password, setPassword] = useState(""); //비밀번호는 빈 값으로 둠! (보통 그렇지 않나;;)
-  const [checkPassword, setCheckPassword] = useState("");
-
-  //빨간 텍스트
-  const [pwWarning, setPwWarning] = useState("");
-  //비밀번호 일치/비일치 상태
-  const [pwState, setPwState] = useState();
-
-  //비밀번호 눈 버튼
-  const [pwVisibleBtn, setPwVisibleBtn] = useState(false);
-  const [checkPwVisibleBtn, setCheckPwVisibleBtn] = useState(false);
-
-  //토스트
-  const { toast, showToast } = useToast();
-  //배경 이미지 (추후에 따로 분리해두면 좋을듯)
-  const [selectedBackground, setSelectedBackground] = useState("GREEN"); //기본값주기?
-  const backgrounds = [
-    {
-      id: 1,
-      name: "GREEN",
-      src: null,
-      className: "background-green",
-      alt: "background image - green",
-    },
-    {
-      id: 2,
-      name: "YELLOW",
-      src: null,
-      className: "background-yellow",
-      alt: "background image - yellow",
-    },
-    {
-      id: 3,
-      name: "BLUE",
-      src: null,
-      className: "background-blue",
-      alt: "background image - blue",
-    },
-    {
-      id: 4,
-      name: "PINK",
-      src: null,
-      className: "background-pink",
-      alt: "background image - pink",
-    },
-    {
-      id: 5,
-      name: "DESK1",
-      src: desk1,
-      className: null,
-      alt: "background image - desk1",
-    },
-    {
-      id: 6,
-      name: "DESK2",
-      src: desk2,
-      className: null,
-      alt: "background image - desk2",
-    },
-    {
-      id: 7,
-      name: "TILE",
-      src: tile,
-      className: null,
-      alt: "background image - tile",
-    },
-    {
-      id: 8,
-      name: "LEAF",
-      src: leaf,
-      className: null,
-      alt: "background image - leaf",
-    },
-  ];
-
-  //비밀번호 확인 입력 시, 비밀번호와 일치하는지 검사하고, 일치하지 않다면 문구 띄우기
-  useEffect(() => {
-    if (password.trim() !== "" && checkPassword.trim() !== "") {
-      if (password !== checkPassword) {
-        setPwWarning("* 비밀번호가 일치하지 않습니다!");
-        setPwState(false);
-      } else if (password === checkPassword) {
-        setPwWarning("비밀번호가 일치합니다.");
-        setPwState(true);
-      }
-    } else {
-      setPwWarning("");
-    }
-  }, [password, checkPassword]);
-
-  //만들기/수정하기 버튼 클릭 시
-  const onHandleSubmit = async () => {
-    //설명 제외 모든 곳에 입력이 되었는지 검사 -> 입력 안된 부분 있으면 토스트 메세지로 띄우기
-    if (
-      nickname.trim() === "" ||
-      title.trim() === "" ||
-      description.trim() === "" ||
-      password.trim() === "" ||
-      checkPassword.trim() === ""
-    ) {
-      showToast("warning", "입력이 필요합니다!");
-      return;
-    }
-    //비밀번호와 비밀번호 확인이 일치한지 검사 -> 다르면 토스트 메세지로 띄우기
-    if (password !== checkPassword) {
-      showToast("warning", "비밀번호가 일치하지 않습니다!");
-      return;
-    }
-
-    //생성/수정 api 보내기
-    let result;
-    // study.id = 8; //테스트용
-    switch (type) {
-      case "create":
-        console.log("스터디 생성 요청 시작");
-        result = await createStudy({
-          title,
-          nickname,
-          description,
-          password,
-          theme: selectedBackground,
-        });
-        console.log("result=>", result);
-
-        break;
-      case "modify":
-        console.log("스터디 수정 요청 시작");
-        result = await updateStudy(study.id, {
-          title,
-          nickname,
-          description,
-          password,
-          theme: selectedBackground,
-        });
-        console.log("result=>", result);
-        break;
-      default:
-        console.log("type이 잘못되었습니다. type=>", type);
-    }
-    showToast("success", "등록되었습니다!");
-    console.log("이 id의 상세 페이지로 이동 =>", result.data.data.id);
-    navigate(`/studies/${result.data.data.id}`); //create,modify모두 스터디 상세 페이지로 이동하도록 한다.
-    //데이터가 등록된 후에 이동해야 하기에, Link가 아닌 navigate를 이용.
-  };
   return (
     <div className={styles.container}>
       <div className={styles.studyFormContainer}>
-        <h2 className={styles.formTitle}>
-          {type === "modify" ? "스터디 수정" : "스터디 만들기"}
-        </h2>
-        <form className={styles.studyForm}>
-          <div className={styles.inputElement}>
-            <label htmlFor="nickname-input" className={styles.label}>
-              닉네임
-            </label>
-            <input
-              type="text"
-              id="nickname-input"
-              value={nickname}
-              onChange={(e) => setNickname(e.target.value)}
-              placeholder="닉네임을 입력해 주세요"
-              className={`${styles.input} ${!nickname && styles.noContent}`}
-            />
-            {!nickname && (
-              <p className={styles.warningText}>* 닉네임을 입력해 주세요</p>
-            )}
-          </div>
-          <div className={styles.inputElement}>
-            <label htmlFor="title-input" className={styles.label}>
-              스터디 이름
-            </label>
-            <input
-              type="text"
-              id="title-input"
-              value={title}
-              onChange={(e) => setTitle(e.target.value)}
-              placeholder="스터디 이름을 입력해 주세요"
-              className={`${styles.input} ${!title && styles.noContent}`}
-            />
-            {!title && (
-              <p className={styles.warningText}>
-                * 스터디 이름을 입력해 주세요
-              </p>
-            )}
-          </div>
-          <div className={styles.inputElement}>
-            <label htmlFor="description-input" className={styles.label}>
-              소개
-            </label>
-            <textarea
-              type="text"
-              id="description-input"
-              value={description}
-              onChange={(e) => setDescription(e.target.value)}
-              placeholder="소개 멘트를 작성해 주세요"
-              className={styles.textarea}
-            />
-          </div>
-          <p className={styles.label}>배경을 선택해주세요</p>
-          <div className={styles.backgrounds}>
-            {backgrounds.map((background) => {
-              return (
-                <div
-                  key={background.id}
-                  onClick={() => {
-                    setSelectedBackground(background.name);
-                  }}
-                  className={styles.background}
-                >
-                  {background.className && (
-                    <div
-                      className={`${styles.colorchip} ${styles[background.className]}`}
-                    />
-                  )}
-                  {background.src && (
-                    <img src={background.src} alt={background.alt} />
-                  )}
-
-                  {selectedBackground === background.name ? (
-                    <div className={styles.pawContainer}>
-                      <div className={`ic paw ${styles.paw}`} />
-                    </div>
-                  ) : (
-                    ""
-                  )}
-                </div>
-              );
-            })}
-          </div>
-          <div className={styles.inputElement}>
-            <label htmlFor="password-input" className={styles.label}>
-              비밀번호
-            </label>
-            <div className={styles.inputContainer}>
-              <input
-                type={pwVisibleBtn ? "text" : "password"}
-                id="password-input"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                placeholder="비밀번호를 입력해 주세요"
-                className={`${styles.input} ${!password && styles.noContent} ${styles.pwInput}`}
-              />
-
-              <button
-                type="button"
-                className={pwVisibleBtn ? "ic visible" : "ic hidden"}
-                onClick={() => {
-                  setPwVisibleBtn(!pwVisibleBtn);
-                }}
-              />
-            </div>
-            {!password && (
-              <p className={styles.warningText}>* 비밀번호를 입력해 주세요</p>
-            )}
-          </div>
-
-          <div className={styles.inputElement}>
-            <label htmlFor="password-check-input" className={styles.label}>
-              비밀번호 확인
-            </label>
-            <div className={styles.inputContainer}>
-              <input
-                type={checkPwVisibleBtn ? "text" : "password"}
-                id="password-check-input"
-                value={checkPassword}
-                onChange={(e) => setCheckPassword(e.target.value)}
-                placeholder="비밀번호를 한번 더 입력해 주세요"
-                className={styles.input}
-              />
-              <button
-                type="button"
-                className={checkPwVisibleBtn ? "ic visible" : "ic hidden"}
-                onClick={() => {
-                  setCheckPwVisibleBtn(!checkPwVisibleBtn);
-                }}
-              />
-            </div>
-
-            <p
-              className={`${pwState && styles.correctText} ${styles.warningText}`}
-            >
-              {pwWarning}
-            </p>
-          </div>
-          <div className={styles.button}>
-            <Button
-              label={type === "modify" ? "수정하기" : "만들기"}
-              onClick={(e) => {
-                e.preventDefault();
-                onHandleSubmit();
-              }}
-            />
-          </div>
-          <div className={styles.toast}>
-            {toast && <Toast type={toast.type} text={toast.text} />}
-          </div>
-        </form>
+        <Title type={type} />
+        <Form type={type} study={study} />
       </div>
     </div>
   );
