+++ src/pages/StudyFormPage/StudyForm.module.css
@@ -17,179 +17,3 @@
 }
 
 /*-----Form------*/
-.formTitle {
-  color: #414141;
-  font-family: Pretendard;
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 800;
-  line-height: normal;
-}
-.studyForm {
-  width: 100%;
-
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 24px;
-}
-/*-----Input------*/
-.inputElement {
-  width: 100%;
-
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 16px;
-}
-.label {
-  color: var(--black, #414141);
-  font-family: Pretendard;
-  font-size: 18px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-}
-.input {
-  padding: 15px 20px;
-
-  width: 100%;
-  height: 48px;
-
-  border-radius: 15px;
-  border: 1px solid var(--gray-300, #ddd);
-  background: #fff;
-}
-.pwInput {
-  padding-right: 50px; /*visible icon 자리를 위해*/
-}
-.textarea {
-  padding: 15px 20px;
-  padding-right: 50px; /*visible icon 자리를 위해*/
-  resize: none; /* 크기 조절 막기 */
-
-  width: 100%;
-  height: 98px;
-
-  border-radius: 15px;
-  border: 1px solid var(--gray-300, #ddd);
-  background: #fff;
-
-  line-height: 1.5;
-}
-.noContent {
-  border: 1px solid var(--error, #c41013);
-}
-.input::placeholder {
-  color: var(--gray-500, #818181);
-  font-family: Pretendard;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
-.inputContainer {
-  width: 100%;
-
-  position: relative;
-
-  display: flex;
-  align-items: center;
-}
-.inputContainer button {
-  width: 20.473px;
-  height: 18.065px;
-
-  position: absolute;
-  right: 20px;
-}
-
-/*-----배경 이미지------*/
-.backgrounds {
-  width: 100%;
-
-  display: flex;
-  gap: 16px;
-  flex-wrap: wrap;
-  align-items: flex-start;
-  align-content: flex-start;
-}
-.background {
-  width: 150px;
-  height: 150px;
-
-  border-radius: 16px;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  background: var(--green, #e1edde);
-  cursor: pointer;
-
-  position: relative;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  overflow: hidden;
-}
-.background img,
-.background .colorchip {
-  width: 100%;
-  height: 100%;
-  object-fit: cover; /* 이미지가 찌그러지지 않게 */
-}
-
-.backgrounds .selected {
-}
-
-.background-green {
-  background: var(--green, #e1edde);
-}
-.background-yellow {
-  background: var(--yellow, #ebedde);
-}
-.background-blue {
-  background: var(--blue, #dee2ed);
-}
-.background-pink {
-  background: var(--pink, #eddee9);
-}
-/*-----부가 텍스트-----*/
-.warningText {
-  color: var(--error, #c41013);
-  font-family: Pretendard;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
-.correctText {
-  color: var(--green-text);
-}
-/*-----아이콘-----*/
-.pawContainer {
-  width: 32px;
-  height: 32px;
-
-  border-radius: 50%;
-  background-color: rgba(65, 65, 65, 0.5);
-
-  position: absolute;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.paw {
-  width: 18px;
-  height: 18px;
-}
-
-/*-----Button------*/
-.button {
-  width: 100%;
-  max-width: 648px;
-  height: 58px;
-}
-/*-----Toast------*/
-.toast {
-  z-index: 1000;
-}
