+++ src/pages/StudyFormPage/components/Backgrounds/Backgrounds.module.css
@@ -1,0 +1,76 @@
+.label {
+  color: var(--black, #414141);
+  font-family: Pretendard;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: normal;
+}
+
+/*-----배경 이미지------*/
+.backgrounds {
+  width: 100%;
+
+  display: flex;
+  gap: 16px;
+  flex-wrap: wrap;
+  align-items: flex-start;
+  align-content: flex-start;
+}
+.background {
+  width: 150px;
+  height: 150px;
+
+  border-radius: 16px;
+  border: 1px solid rgba(0, 0, 0, 0.1);
+  background: var(--green, #e1edde);
+  cursor: pointer;
+
+  position: relative;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  overflow: hidden;
+}
+.background img,
+.background .colorchip {
+  width: 100%;
+  height: 100%;
+  object-fit: cover; /* 이미지가 찌그러지지 않게 */
+}
+
+.backgrounds .selected {
+}
+
+.background-green {
+  background: var(--green, #e1edde);
+}
+.background-yellow {
+  background: var(--yellow, #ebedde);
+}
+.background-blue {
+  background: var(--blue, #dee2ed);
+}
+.background-pink {
+  background: var(--pink, #eddee9);
+}
+
+/*-----아이콘-----*/
+.pawContainer {
+  width: 32px;
+  height: 32px;
+
+  border-radius: 50%;
+  background-color: rgba(65, 65, 65, 0.5);
+
+  position: absolute;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.paw {
+  width: 18px;
+  height: 18px;
+}
