+++ src/pages/StudyFormPage/components/Input/Input.module.css
@@ -1,0 +1,88 @@
+/*-----Input------*/
+.inputElement {
+  width: 100%;
+
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 16px;
+}
+.label {
+  color: var(--black, #414141);
+  font-family: Pretendard;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: normal;
+}
+.input {
+  padding: 15px 20px;
+
+  width: 100%;
+  height: 48px;
+
+  border-radius: 15px;
+  border: 1px solid var(--gray-300, #ddd);
+  background: #fff;
+}
+.pwInput {
+  padding-right: 50px; /*visible icon 자리를 위해*/
+}
+.textarea {
+  padding: 15px 20px;
+  padding-right: 50px; /*visible icon 자리를 위해*/
+  resize: none; /* 크기 조절 막기 */
+
+  width: 100%;
+  height: 98px;
+
+  border-radius: 15px;
+  border: 1px solid var(--gray-300, #ddd);
+  background: #fff;
+
+  line-height: 1.5;
+}
+.noContent {
+  border: 1px solid var(--error, #c41013);
+}
+.input::placeholder {
+  color: var(--gray-500, #818181);
+  font-family: Pretendard;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+}
+.inputContainer {
+  width: 100%;
+
+  position: relative;
+
+  display: flex;
+  align-items: center;
+}
+.inputContainer button {
+  width: 20.473px;
+  height: 18.065px;
+
+  position: absolute;
+  right: 20px;
+}
+
+/*-----부가 텍스트-----*/
+.warningText {
+  color: var(--error, #c41013);
+  font-family: Pretendard;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+}
+.correctText {
+  color: var(--green-text);
+  font-family: Pretendard;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+}
