+++ src/pages/StudyFormPage/components/Title/Title.module.css
@@ -1,0 +1,8 @@
+.formTitle {
+  color: #414141;
+  font-family: Pretendard;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 800;
+  line-height: normal;
+}
